fix(frontend): seperated Connection Component into it's own file and resolve hydration error
